Remove submit & admin form. Add simple edit form for onsite editing when needed.
